% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycobacteriumAnalyze.R
\name{get_data}
\alias{get_data}
\title{import clusterProfiler
import UniProt.ws
Reads in data from the disk. This data should be in CSV format with the
first column as the protein names and the second column as the gene names.}
\usage{
get_data(filename)
}
\arguments{
\item{filename}{The filename of the CSV file you want to read in.}
}
\value{
A dataframe with two columns. The first labeled "protein" the second "gene"
}
\description{
import clusterProfiler
import UniProt.ws
Reads in data from the disk. This data should be in CSV format with the
first column as the protein names and the second column as the gene names.
}
