% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycobacteriumAnalyze.R
\name{get_KEGG_enrichment}
\alias{get_KEGG_enrichment}
\title{Calculates and returns results for KEGG enrichment given a species
and corresponding UniProt IDs.}
\usage{
get_KEGG_enrichment(uniprot_IDs, species, pvalueCutoff = 0.05)
}
\arguments{
\item{uniprot_IDs}{The corresponding UniProt IDs you want to test}

\item{species}{the species you are analyzing (either 'avium' or 'abscessus')}

\item{pvalueCutoff}{The p-value threshold that is considered significant (optional, default=0.05)}
}
\value{
A results object from clusterProfiler
}
\description{
Calculates and returns results for KEGG enrichment given a species
and corresponding UniProt IDs.
}
