% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycobacteriumAnalyze.R
\name{get_uniprot_IDs}
\alias{get_uniprot_IDs}
\title{Retrieves the UniProt IDs for a given species that are compatible for 
KEGG enrichment.}
\usage{
get_uniprot_IDs(data, species)
}
\arguments{
\item{data}{The orignal protein data composed of a 'gene' column and a 'protein' column}

\item{species}{The species you are analyzing (either 'avium' or 'abscessus')}
}
\description{
Retrieves the UniProt IDs for a given species that are compatible for 
KEGG enrichment.
}
