% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycobacteriumAnalyze.R
\name{write_KEGG_enrichment}
\alias{write_KEGG_enrichment}
\title{Writes the results from the KEGG enrichment object to disk as a comma
separated value format file.}
\usage{
write_KEGG_enrichment(KEGG_enrichment_object, filename)
}
\arguments{
\item{KEGG_enrichment_object}{The clusterProfiler KEGG enrichment results}

\item{filename}{The location where the file should be written}
}
\value{
No return value
}
\description{
Writes the results from the KEGG enrichment object to disk as a comma
separated value format file.
}
